# api/openapi.yaml
basePath: /
definitions:
...
  domain.AnalyticsPeriod:
    enum:
    - day
    - week
    - month
    - year
    type: string
    x-enum-varnames:
    - AnalyticsDay
    - AnalyticsWeek
    - AnalyticsMonth
    - AnalyticsYear
  domain.AuditEntry:
    properties:
      action:
        description: Action is the event type that produced the entry.
        type: string
      actorID:
        description: ActorID is the user who caused the event, when known.
        type: string
      eventID:
        description: EventID is the source event, used to make replay idempotent.
        type: string
      metadata:
        additionalProperties:
          type: string
        description: Metadata carries additional non-indexed context.
        type: object
      occurredAt:
        description: OccurredAt is when the command produced the event.
        type: string
      resourceID:
        description: ResourceID is the aggregate the event belongs to.
        type: string
      resourceType:
        description: ResourceType is the aggregate type the event belongs to.
        type: string
    type: object
...
  domain.BudgetSuggestion:
    properties:
      budgetID:
        type: string
      categoryID:
        type: string
      currency:
        type: string
      limitMinor:
        type: integer
      message:
        type: string
      percentUsed:
        type: integer
      spentMinor:
        type: integer
    type: object
...
  domain.ExpenseRollup:
    properties:
      amountMinor:
        description: AmountMinor is the aggregate in currency minor units.
        type: integer
      categoryID:
        description: CategoryID optionally identifies the aggregated category.
        type: string
      currency:
        description: Currency is the ISO 4217 currency code.
        type: string
      expenseCount:
        description: ExpenseCount is the number of included expenses.
        type: integer
      ownerID:
        description: OwnerID identifies the user whose expenses were aggregated.
        type: string
      period:
        allOf:
        - $ref: '#/definitions/domain.AnalyticsPeriod'
        description: Period identifies the rollup window size.
      periodStart:
        description: PeriodStart is the inclusive start of the aggregation window.
        type: string
      tagID:
        description: TagID optionally identifies the aggregated tag.
        type: string
    type: object
...
  handler.assistantRequest:
    properties:
      message:
        type: string
    type: object
...
info:
  contact: {}
  description: API for personal cost tracking.
  title: BillPiggy API
  version: "1.0"
paths:
...
  /billpiggy/api/v1/analytics/expenses:
    get:
      parameters:
      - description: day, week, month, or year
        in: query
        name: period
        required: true
        type: string
      - description: RFC3339 inclusive start
        in: query
        name: from
        required: true
        type: string
      - description: RFC3339 inclusive end
        in: query
        name: to
        required: true
        type: string
      - description: Category ID
        in: query
        name: category_id
        type: string
      - collectionFormat: csv
        description: Tag IDs
        in: query
        items:
          type: string
        name: tag_id
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.ExpenseRollup'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get expense analytics
      tags:
      - analytics
  /billpiggy/api/v1/analytics/suggestions:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.BudgetSuggestion'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get budget suggestions
      tags:
      - analytics
  /billpiggy/api/v1/assistant/chat:
    post:
      consumes:
      - application/json
      parameters:
      - description: Assistant message
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.assistantRequest'
      produces:
      - text/event-stream
      responses:
        "200":
          description: 'SSE events: message.started, message.delta, message.completed,
            message.error'
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Stream an assistant response
      tags:
      - assistant
  /billpiggy/api/v1/audit:
    get:
      parameters:
      - description: Actor user ID
        in: query
        name: actor_id
        type: string
      - description: Aggregate type, e.g. expense, budget, user
        in: query
        name: resource_type
        type: string
      - description: Aggregate ID
        in: query
        name: resource_id
        type: string
      - description: Event type, e.g. expense_added
        in: query
        name: action
        type: string
      - description: RFC3339 inclusive start
        in: query
        name: from
        type: string
      - description: RFC3339 exclusive end
        in: query
        name: to
        type: string
      - description: Page size, default 50, max 200
        in: query
        name: limit
        type: integer
      - description: Page offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.AuditEntry'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: List audit entries
      tags:
      - administration
...
securityDefinitions:
  ApiKeyAuth:
    description: Type "Bearer" followed by a space and the access token from POST
      /auth/login or POST /auth/refresh.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"